Allow adding decorative sprites
You can add such a sprite in Tiled by adding a rectangle with the size that you want, and giving it the name of the .png file in the 'decorations' folder in resources that you want to show there.
These objects are meant to be static decorations in the levels. Just to give the levels somewhat more flare.